Improve provider shutdown, networking, and payment compatibility
Add graceful shutdown and termination notices, update relay and VPN
handling, and preserve payment precision through database updates.

+++ core/payment/src/api/debit_notes.rs
@@ -2,7 +2,6 @@ use std::borrow::Cow;
 // Extrnal crates
 use actix_web::web::{get, post, Data, Json, Path, Query};
 use actix_web::{HttpResponse, Scope};
-use bigdecimal::BigDecimal;
 use serde_json::value::Value::Null;
 use std::time::Instant;
 // Workspace uses
@@ -423,7 +422,7 @@ async fn accept_debit_note(
         return response::bad_request(&msg);
     }
 
-    if amount_to_pay > BigDecimal::from(0) {
+    if amount_to_pay > 0 {
         match db
             .as_dao::<AllocationDao>()
             .spend_from_allocation_transaction(SpendFromAllocationArgs {
